python/requirements.txt
kafka-python
VaderSentiment
